更新 build.sh

+++ nn6000v2/scripts/build.sh
@@ -127,8 +127,12 @@ if [[ $Build_Mod == "debug" ]]; then
 fi
 
 TARGET_DIR="$BASE_PATH/../$BUILD_DIR/bin/targets"
-if [[ -d $TARGET_DIR ]]; then
-    find "$TARGET_DIR" -type f \( -name "*.bin" -o -name "*.manifest" -o -name "*efi.img.gz" -o -name "*.itb" -o -name "*.fip" -o -name "*.ubi" -o -name "*rootfs.tar.gz" \) -exec rm -f {} +
+
+# Only remove existing firmware files if not building nowifi version
+if [[ "$Version_Tag" != "nowifi" ]]; then
+    if [[ -d $TARGET_DIR ]]; then
+        find "$TARGET_DIR" -type f \( -name "*.bin" -o -name "*.manifest" -o -name "*efi.img.gz" -o -name "*.itb" -o -name "*.fip" -o -name "*.ubi" -o -name "*rootfs.tar.gz" \) -exec rm -f {} +
+    fi
 fi
 
 make download -j$(($(nproc) * 2))
